# docs/04-api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: LinkFlow API
  description: |
    LinkFlow is a high-performance workflow automation platform.
    This API serves as the control plane for managing workspaces, workflows,
    credentials, and executions.
  version: 1.0.0
  contact:
    name: LinkFlow Support
    email: support@linkflow.io
  license:
    name: Proprietary
    url: https://linkflow.io/terms
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.linkflow.io/api/v1
    description: Production
...
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, email, password, password_confirmation]
              properties:
                token:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                password_confirmation:
                  type: string
                  format: password
      responses:
        '200':
          description: Password reset successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthenticated.
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
  schemas:
...
    RefreshTokenResponse:
      type: object
      required: [message, access_token, refresh_token, token_type, expires_in]
      properties:
        message:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Seconds until access token expiry
...